requirements.txt
alabaster==0.7.12
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
enum34==1.1.6
filelock==3.0.10
flake8==3.7.7
funcsigs==1.0.2
greenlet==0.4.15
idna==2.8
imagesize==1.1.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.1.0
llvmlite==0.28.0
lockfile==0.12.2
luigi==2.8.3
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.0.0
mpmath==1.1.0
msgpack==0.6.1
nbconvert==5.5.0
nbformat==4.4.0
neovim==0.3.1
notebook==5.7.8
numba==0.43.1
numpy==1.16.3
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.9.0
prince==0.6.2
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.3.1
pynvim==0.3.2
pyparsing==2.4.0
pyrsistent==0.15.1
pytest==4.4.1
pytest-cov==2.6.1
python-daemon==2.1.2
python-dateutil==2.7.5
pytorch-ignite==0.2.0
pytz==2019.1
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
snakeviz==2.0.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.0
src==0.1.0
sympy==1.4
terminado==0.8.2
testpath==0.4.2
toml==0.10.0
torch==1.0.1.post2
torchsummary==1.5.1
torchvision==0.2.2.post3
tornado==4.5.3
tox==3.9.0
tqdm==4.31.1
traitlets==4.3.2
umap==0.1.1
umap-learn==0.3.8
urllib3==1.24.2
virtualenv==16.5.0
wcwidth==0.1.7
webencodings==0.5.1
-e git+git@github.com:arnor-sigurdsson/wgan-gp_human_origins.git@e62412f548d1a6ded2661658e95eb88909599f64#egg=wgan_gp_human_origins
widgetsnbextension==3.4.2
